# local-shared-data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/public/artists/{slug}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Public Artist",
        "operationId": "get_public_artist",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicArtistRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/public/artists/{slug}/portfolio": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Public Portfolio",
        "operationId": "get_public_portfolio",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortfolioImageRead"
                  },
                  "title": "Response Public-Get Public Portfolio"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/public/artists/{slug}/products": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Public Products",
        "operationId": "get_public_products",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductRead"
                  },
                  "title": "Response Public-Get Public Products"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/public/artists/{slug}/checkout": {
      "post": {
        "tags": [
          "public"
        ],
        "summary": "Checkout",
        "operationId": "checkout",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/public/artists/{slug}/flash": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Public Flash",
        "operationId": "get_public_flash",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlashPieceRead"
                  },
                  "title": "Response Public-Get Public Flash"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/public/artists/{slug}/slots": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Public Slots",
        "operationId": "get_public_slots",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SlotRead"
                  },
                  "title": "Response Public-Get Public Slots"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/public/artists/{slug}/book": {
      "post": {
        "tags": [
          "public"
        ],
        "summary": "Book",
        "operationId": "book",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/uploads/image": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Upload Image",
        "operationId": "upload_image",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploads-upload_image"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_uploads-upload_image": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_uploads-upload_image"
      },
...
      "BookingRequest": {
        "properties": {
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At"
          },
          "end_at": {
            "type": "string",
            "format": "date-time",
            "title": "End At"
          },
          "client_name": {
            "type": "string",
            "title": "Client Name"
          },
          "client_email": {
            "type": "string",
            "format": "email",
            "title": "Client Email"
          },
          "client_phone": {
            "type": "string",
            "title": "Client Phone"
          },
          "size": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "placement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placement"
          },
          "color_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ColorType"
              },
              {
                "type": "null"
              }
            ]
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "accommodations_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accommodations Notes"
          },
          "has_guests": {
            "type": "boolean",
            "title": "Has Guests",
            "default": false
          },
          "guests_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guests Notes"
          },
          "payment_method_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Method Id"
          },
          "age_confirmed": {
            "type": "boolean",
            "title": "Age Confirmed",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "start_at",
          "end_at",
          "client_name",
          "client_email",
          "client_phone"
        ],
        "title": "BookingRequest"
      },
      "BookingResult": {
        "properties": {
          "appointment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Appointment Id"
          },
          "status": {
            "$ref": "#/components/schemas/AppointmentStatus"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "appointment_id",
          "status",
          "message"
        ],
        "title": "BookingResult"
      },
...
      "CheckoutRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CheckoutItem"
            },
            "type": "array",
            "title": "Items"
          },
          "customer_name": {
            "type": "string",
            "title": "Customer Name"
          },
          "customer_email": {
            "type": "string",
            "format": "email",
            "title": "Customer Email"
          },
          "customer_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Phone"
          },
          "shipping_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipping Address"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "accepted_terms": {
            "type": "boolean",
            "title": "Accepted Terms",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "items",
          "customer_name",
          "customer_email"
        ],
        "title": "CheckoutRequest"
      },
      "CheckoutResult": {
        "properties": {
          "order_id": {
            "type": "string",
            "format": "uuid",
            "title": "Order Id"
          },
          "total_cents": {
            "type": "integer",
            "title": "Total Cents"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "total_cents",
          "status",
          "message"
        ],
        "title": "CheckoutResult"
      },
...
      "FlashPieceRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "artist_id": {
            "type": "string",
            "format": "uuid",
            "title": "Artist Id"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price_min_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Min Cents"
          },
          "price_max_cents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Max Cents"
          },
          "price_plus": {
            "type": "boolean",
            "title": "Price Plus"
          },
          "size_min": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Min"
          },
          "size_plus": {
            "type": "boolean",
            "title": "Size Plus"
          },
          "ask_about": {
            "type": "boolean",
            "title": "Ask About"
          },
          "status": {
            "$ref": "#/components/schemas/FlashStatus"
          },
          "event_tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Tag"
          }
        },
        "type": "object",
        "required": [
          "id",
          "artist_id",
          "image_url",
          "price_plus",
          "size_plus",
          "ask_about",
          "status"
        ],
        "title": "FlashPieceRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PortfolioImageRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "artist_id": {
            "type": "string",
            "format": "uuid",
            "title": "Artist Id"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "is_healed": {
            "type": "boolean",
            "title": "Is Healed"
          },
          "tattoo_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tattoo Date"
          }
        },
        "type": "object",
        "required": [
          "id",
          "artist_id",
          "image_url",
          "sort_order",
          "is_healed"
        ],
        "title": "PortfolioImageRead"
      },
...
      "ProductRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "artist_id": {
            "type": "string",
            "format": "uuid",
            "title": "Artist Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price_cents": {
            "type": "integer",
            "title": "Price Cents"
          },
          "images": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Images"
          },
          "inventory": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inventory"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "artist_id",
          "title",
          "price_cents",
          "active"
        ],
        "title": "ProductRead"
      },
...
      "PublicArtistRead": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "instagram_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instagram Url"
          },
          "deposit_default_cents": {
            "type": "integer",
            "title": "Deposit Default Cents"
          },
          "accepting_bookings": {
            "type": "boolean",
            "title": "Accepting Bookings"
          },
          "require_review_before_confirm": {
            "type": "boolean",
            "title": "Require Review Before Confirm"
          },
          "theme": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ThemeSettingsRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "announcement_banner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Announcement Banner"
          },
          "modules": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/BookingModule"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modules"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "display_name",
          "deposit_default_cents",
          "accepting_bookings",
          "require_review_before_confirm"
        ],
        "title": "PublicArtistRead"
      },
      "SlotRead": {
        "properties": {
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At"
          },
          "end_at": {
            "type": "string",
            "format": "date-time",
            "title": "End At"
          }
        },
        "type": "object",
        "required": [
          "start_at",
          "end_at"
        ],
        "title": "SlotRead"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  }
}